requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.0.0
plotly>=5.15.0
altair>=5.0.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0
# Advanced Analytics & ML
scipy>=1.11.0
statsmodels>=0.14.0
prophet>=1.1.4
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
# Time Series & Forecasting
pmdarima>=2.0.4
arch>=6.2.0
# Advanced Visualization
bokeh>=3.2.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
# Data Processing
pyarrow>=14.0.0
fastparquet>=2023.10.0
# API & Web
flask>=2.3.0
flask-cors>=4.0.0
# Utilities
tqdm>=4.66.0
joblib>=1.3.0